Algunos cambios
