+++ src/main/java/BBDD/ModelosBD/ClientePropietarioEntity.java
@@ -1,4 +1,4 @@
-package BBDD;
+package BBDD.ModelosBD;
 
 import java.util.ArrayList;
 
@@ -6,13 +6,13 @@ public class ClientePropietarioEntity extends ClienteEntity {
     private ArrayList<InmuebleEntity> inmuebles = new ArrayList<>();
 
 
-    public ClientePropietarioEntity(int cedula, String nombre, String apellido, int edad, int diaNacimiento, int mesNacimiento,
+    /*public ClientePropietarioEntity(int cedula, String nombre, String apellido, int edad, int diaNacimiento, int mesNacimiento,
                                     int anhoNacimiento, String email, int telefono, String trabajo, String parroquiaDireccion,
                                     String municipioDireccion, String estadoDireccion, String ciudadDireccion, String direccion,
                                     float precioInicial) {
         super(cedula, nombre, apellido, edad, diaNacimiento, mesNacimiento, anhoNacimiento, email, telefono, trabajo,
                 parroquiaDireccion, municipioDireccion, estadoDireccion, ciudadDireccion, direccion);
-    }
+    }*/
 
     public void addInmueble(InmuebleEntity inmueble) {
         this.inmuebles.add(inmueble);
